Self_Attention_Arbitrary_Style_Transfer/requirements.txt
numpy==1.19.2
Pillow==8.1.0
scipy==1.6.0
tqdm==4.56.0